first class groupings
